+++ contrib/pai_vscode/src/pai/container/storage/sambaTreeItem.ts
@@ -58,7 +58,7 @@ export class SambaRootNode extends StorageTreeNode {
         await PathBaseStorageManager.uploadFolders(this);
     }
 
-    public async createFolder(): Promise<void> {
-        await PathBaseStorageManager.createFolder(this);
+    public async createFolder(folder?: string): Promise<void> {
+        await PathBaseStorageManager.createFolder(this, folder);
     }
 }