+++ src/kube-runtime/src/plugins/ssh/init.py
@@ -22,7 +22,7 @@ import sys
 
 sys.path.append(
     os.path.join(os.path.dirname(os.path.abspath(__file__)), "../.."))
-from plugins.plugin_utils import plugin_init, PluginHelper  #pylint: disable=wrong-import-position
+from plugins.plugin_utils import plugin_init, PluginHelper, try_to_install_by_cache  #pylint: disable=wrong-import-position
 
 LOGGER = logging.getLogger(__name__)
 
@@ -37,7 +37,11 @@ def main():
         LOGGER.info("Ssh plugin parameters is empty, ignore this")
         return
 
-    if "jobssh" in parameters:
+    gang_allocation = os.environ.get("GANG_ALLOCATION", "true")
+    if gang_allocation == "false":
+        LOGGER.warning("Job ssh is conflict with gang allocation, set job ssh to false")
+        jobssh = "false"
+    elif "jobssh" in parameters:
         jobssh = str(parameters["jobssh"]).lower()
     else:
         jobssh = "false"
@@ -49,10 +53,18 @@ def main():
             cmd_params.append("\'{}\'".format(parameters["userssh"]["value"]))
 
     # write call to real executable script
-    command = [
-        "{}/sshd.sh {}\n".format(os.path.dirname(os.path.abspath(__file__)),
-                                 " ".join(cmd_params))
-    ]
+    command = []
+    if len(cmd_params) == 1 and cmd_params[0] == "false":
+        LOGGER.info("Skip sshd script since neither jobssh or userssh is set")
+    else:
+        command = [
+            try_to_install_by_cache("ssh", fallback_cmds=[
+                "apt-get update",
+                "apt-get install -y openssh-client openssh-server",
+            ]),
+            "{}/sshd.sh {}\n".format(os.path.dirname(os.path.abspath(__file__)),
+                                     " ".join(cmd_params))
+        ]
 
     # ssh barrier
     if jobssh == "true" and "sshbarrier" in parameters and str(